Details - Fixed AdjustFontSize
