+++ .idea/misc.xml
@@ -7,4 +7,11 @@
   <component name="ProjectType">
     <option name="id" value="Android" />
   </component>
+  <component name="VisualizationToolProject">
+    <option name="state">
+      <ProjectState>
+        <option name="scale" value="0.01965901183020181" />
+      </ProjectState>
+    </option>
+  </component>
 </project>
