+++ app/src/main/res/drawable/rounded_bottom_toolbar.xml
@@ -2,11 +2,22 @@
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:bottom="4dp">
         <shape android:shape="rectangle">
-            <solid android:color="@color/main_gray"/>
-            <corners android:topLeftRadius="0dp"
-                android:topRightRadius="0dp"
+            <solid android:color="@color/main_gray" />
+            <corners
                 android:bottomLeftRadius="16dp"
-                android:bottomRightRadius="16dp"/>
+                android:bottomRightRadius="16dp"
+                android:topLeftRadius="0dp"
+                android:topRightRadius="0dp" />
+            <padding
+                android:bottom="24dp"
+                android:left="16dp"
+                android:right="16dp"
+                android:top="10dp" />
+        </shape>
+    </item>
+    <item android:top="24dp">
+        <shape android:shape="rectangle">
+            <solid android:color="@android:color/transparent" />
         </shape>
     </item>
 </layer-list>
