+++ app/src/main/res/layout/activity_main.xml
@@ -5,12 +5,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".ui.MainActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/customToolbar"
         android:layout_width="0dp"
-        android:layout_height="70dp"
+        android:layout_height="107dp"
         android:background="@drawable/rounded_bottom_toolbar"
         android:padding="5dp"
         app:layout_constraintBottom_toTopOf="@+id/scrollView6"
@@ -20,8 +20,8 @@
 
         <ImageView
             android:id="@+id/toolbarImage"
-            android:layout_width="200dp"
-            android:layout_height="wrap_content"
+            android:layout_width="178dp"
+            android:layout_height="31dp"
             android:layout_gravity="center"
             android:contentDescription="@string/logo_description"
             android:scaleType="centerCrop"
@@ -52,7 +52,8 @@
                 android:layout_height="wrap_content"
                 android:contentDescription="@string/h_scrollview_description1"
                 android:scrollbars="none"
-                app:layout_constraintHeight_percent="0.2">
+                app:layout_constraintHeight_percent="0.2"
+                app:layout_goneMarginTop="16dp">
 
                 <LinearLayout
                     android:id="@+id/cardContainer1"
@@ -69,6 +70,7 @@
                 android:contentDescription="@string/h_scrollview_description2"
                 android:scrollbars="none"
                 app:layout_constraintHeight_percent="0.2"
+                app:layout_goneMarginTop="16dp"
                 tools:layout_editor_absoluteY="32dp">
 
                 <LinearLayout
@@ -86,6 +88,7 @@
                 android:contentDescription="@string/h_scrollview_description3"
                 android:scrollbars="none"
                 app:layout_constraintHeight_percent="0.2"
+                app:layout_goneMarginTop="16dp"
                 tools:layout_editor_absoluteX="190dp"
                 tools:layout_editor_absoluteY="64dp">
 
@@ -104,6 +107,7 @@
                 android:contentDescription="@string/h_scrollview_description4"
                 android:scrollbars="none"
                 app:layout_constraintHeight_percent="0.2"
+                app:layout_goneMarginTop="16dp"
                 tools:layout_editor_absoluteX="190dp"
                 tools:layout_editor_absoluteY="96dp">
 
@@ -122,6 +126,7 @@
                 android:contentDescription="@string/h_scrollview_description5"
                 android:scrollbars="none"
                 app:layout_constraintHeight_percent="0.2"
+                app:layout_goneMarginTop="16dp"
                 tools:layout_editor_absoluteX="190dp"
                 tools:layout_editor_absoluteY="128dp">
