+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Tech Challenge Kotlin</string>
+    <string name="app_name">AvantSoft</string>
 
     <!--    Components Tag-->
     <string name="text_name_title">\Name</string>
